+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Treat a doubled `%%` as an escaped percent sign in environment expressions
+
 ## 0.20.2 (2026-09-13)
 
 - Recognize Stimulus controllers registered manually on a Stimulus application in asset scripts